del selenoidDriverProvider
